+++ src/main/java/com/github/cc3002/finalreality/model/character/player/Engineer.java
@@ -1,6 +1,8 @@
 package com.github.cc3002.finalreality.model.character.player;
 
 import com.github.cc3002.finalreality.model.character.ICharacter;
+import com.github.cc3002.finalreality.model.weapon.Axe;
+import com.github.cc3002.finalreality.model.weapon.Bow;
 import org.jetbrains.annotations.NotNull;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
@@ -31,6 +33,28 @@ public class Engineer extends AbstractPlayerCharacter {
                     @NotNull int hp,
                     @NotNull int defense) {
         super(name, turnsQueue, hp, defense);
+    }
+
+    /**
+     * equips a axe to the engineer.
+     * A engineer class can only equip an axe or a bow. Hence, we create equipAxe, and
+     * equipBow to make sure the engineer can only equip this types of weapons.
+     */
+    public void equipAxe(Axe axe){
+        if (this.IsAlive()){
+            this.setEquippedWeapon(axe);
+        }
+    }
+
+    /**
+     * equips a bow to the engineer.
+     * A engineer class can only equip an axe or a bow. Hence, we create equipAxe, and
+     * equipBow to make sure the engineer can only equip this types of weapons.
+     */
+    public void equipBow(Bow bow){
+        if (this.IsAlive()){
+            this.setEquippedWeapon(bow);
+        }
     }
 
     /**