+++ src/test/java/com/github/cc3002/finalreality/model/character/EnemyTest.java
@@ -1,6 +1,7 @@
 package com.github.cc3002.finalreality.model.character;
 
 import com.github.cc3002.finalreality.model.character.player.Knight;
+import com.github.cc3002.finalreality.model.weapon.Sword;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import java.util.concurrent.BlockingQueue;
@@ -20,7 +21,7 @@ public class EnemyTest{
    */
   protected BlockingQueue<ICharacter> turns = new LinkedBlockingQueue<>();
   private static final String ENEMY_NAME = "Goblin";
-  private Enemy testEnemy;
+  private Enemy testEnemy = new Enemy(ENEMY_NAME, turns, 10,1, 1, 1);
 
   /**
    * This method tests the waitTurn() method if the character is a enemy.
@@ -28,7 +29,6 @@ public class EnemyTest{
    */
   @Test
   void waitTurnTest() {
-    testEnemy = new Enemy(ENEMY_NAME, turns, 10,1, 1, 1);
     Assertions.assertTrue(turns.isEmpty());
     testEnemy.waitTurn();
     try {
@@ -74,4 +74,45 @@ public class EnemyTest{
     assertFalse(expectedEnemy.equals(expectedKnight));
   }
 
+  /**
+   * attackTo() tests.
+   */
+  @Test
+  void attackToTest(){
+    Enemy enemyTest1 = new Enemy("Mimic", turns, 10,50, 3, 10);
+    Enemy enemyTest2 = new Enemy("Iudex", turns, 10,0, 6, 20);
+    Knight knightTest1 = new Knight("Anri", turns, 25, 5);
+    Knight knightTest2 = new Knight("Horace", turns, 0, 4);
+
+    /**
+     * attackTo() if the player is alive, and if the damage is lower than its life.
+     */
+    enemyTest1.attackTo(knightTest1);
+    assertEquals(20, knightTest1.getHP());
+    assertTrue(knightTest1.IsAlive());
+
+    /**
+     * attackTo() if the player is alive, and if the damage is greater than its life.
+     */
+    knightTest1.setHP(4);
+    assertTrue(knightTest1.IsAlive());
+    enemyTest1.attackTo(knightTest1);
+    assertEquals(0, knightTest1.getHP());
+    assertFalse(knightTest1.IsAlive());
+
+    /**
+     * attackTo() if the player is dead.
+     */
+    enemyTest1.attackTo(knightTest2);
+    assertEquals(0, knightTest2.getHP());
+    assertFalse(knightTest2.IsAlive());
+
+    /**
+     * attackTo() if the enemy is dead.
+     */
+    knightTest1.setHP(10);
+    enemyTest2.attackTo(knightTest1);
+    assertEquals(10, knightTest1.getHP());
+    assertTrue(knightTest1.IsAlive());
+  }
 }